drone/requirements.txt
numpy>=1.21.0
pyrtlsdr>=0.2.93
dronekit>=2.9.1
pyserial>=3.5
scipy>=1.7.0
matplotlib>=3.4.0 